Take care of logger
