+++ packages/extensions/core/src/lib/pipeline/plugins/transformer/transformer.ts
@@ -3,12 +3,14 @@ import { createPerFilePlugin, PipelinePlugin } from "../../common";
 import { Manipulator } from "./manipulation";
 import { Channel } from "../../../message";
 import { evalDirectiveTransform } from "./eval";
+import { resolveDirectives } from "@autorest/configuration";
 
 /* @internal */
 export function createGraphTransformerPlugin(): PipelinePlugin {
-  return async (config, input, sink) => {
+  return async (context, input, sink) => {
     // object transforms must have a where clause and a transform
-    const directives = config.resolveDirectives(
+    const directives = resolveDirectives(
+      context.config,
       (x) => x.from.length > 0 && x.transform.length > 0 && x.where.length > 0,
     ); // && (!!x.where && x.where.length > 0)
 
@@ -40,14 +42,14 @@ export function createGraphTransformerPlugin(): PipelinePlugin {
                 // find the target nodes in the document
                 const targets = selectNodes(contents, where);
                 if (targets.length > 0) {
-                  config.Message({
+                  context.Message({
                     Channel: Channel.Debug,
                     Text: `Running object transform '${directive.from}/${directive.reason}' on ${inputHandle.key} `,
                   });
 
                   for (const { path, value, parent } of targets) {
                     const output = evalDirectiveTransform(transform, {
-                      config: config,
+                      config: context,
                       value,
                       parent: parent,
                       doc: contents,
