+++ packages/libs/configuration/src/directive.ts
@@ -1,4 +1,8 @@
+import { createSandbox } from "@azure-tools/datastore";
+import { AutorestConfiguration } from "./autorest-configuration";
 import { arrayOf } from "./utils";
+
+const safeEval = createSandbox();
 
 export interface Directive {
   "from"?: string[] | string;
@@ -34,3 +38,49 @@ export class ResolvedDirective {
     this.test = arrayOf(directive["test"]);
   }
 }
+
+/**
+ * Returns list of ResolvedDirective matching the given predicate.
+ * @param config Configuration containing directives.
+ * @param predicate Optional filter condition.
+ */
+export const resolveDirectives = (
+  config: AutorestConfiguration,
+  predicate?: (each: ResolvedDirective) => boolean,
+): ResolvedDirective[] => {
+  // optionally filter by predicate.
+  const plainDirectives = arrayOf<Directive>(config["directive"]);
+
+  const declarations = config["declare-directive"] || {};
+
+  const expandDirective = (dir: Directive): Directive[] => {
+    const makro = Object.keys(dir).filter((m) => declarations[m])[0];
+    if (!makro) {
+      return [dir]; // nothing to expand
+    }
+
+    // prepare directive
+    let parameters: string[] = (<any>dir)[makro];
+    if (!Array.isArray(parameters)) {
+      parameters = [parameters];
+    }
+    dir = { ...dir };
+    delete (<any>dir)[makro];
+    // call makro
+    const makroResults: any = parameters.flatMap((parameter) => {
+      const result = safeEval(declarations[makro], { $: parameter, $context: dir });
+      return Array.isArray(result) ? result : [result];
+    });
+    return makroResults.flatMap((result: any) => expandDirective({ ...result, ...dir }));
+  };
+
+  // makro expansion
+  if (predicate) {
+    return plainDirectives
+      .flatMap(expandDirective)
+      .map((each) => new ResolvedDirective(each))
+      .filter(predicate);
+  }
+  return plainDirectives.flatMap(expandDirective).map((each) => new ResolvedDirective(each));
+  // return From(plainDirectives).SelectMany(expandDirective).Select(each => new StaticDirectiveView(each)).ToArray();
+};
